# backend/app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HMS API Testing Sandbox",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/pay_bill": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Inventory - /pay_bill",
        "responses": {
          "200": {
            "description": "Successful Operation"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "role": "string_value",
                "institute_id": 123,
                "visit_id": 123,
                "selected_labs": "string_value",
                "selected_medicines": "string_value",
                "success": "string_value",
                "invoice_no": "string_value",
                "POST": "string_value"
              }
            }
          }
        }
      }
    },
    "/cancel_bill": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Inventory - /cancel_bill",
        "responses": {
          "200": {
            "description": "Successful Operation"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "role": "string_value",
                "institute_id": 123,
                "visit_id": 123,
                "success": "string_value",
                "POST": "string_value"
              }
            }
          }
        }
      }
    },
    "/medical_store/bills": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "summary": "Inventory - /medical_store/bills",
        "responses": {
          "200": {
            "description": "Successful Operation"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/medical_store/bills/stats": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "summary": "Inventory - /medical_store/bills/stats",
        "responses": {
          "200": {
            "description": "Successful Operation"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/inventory/add": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Inventory - /inventory/add",
        "responses": {
          "200": {
            "description": "Successful Operation"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "role": "string_value",
                "item_name": "string_value",
                "unit": "string_value",
                "unit_detail": "string_value",
                "item_no": "string_value",
                "sale_rate": "string_value",
                "hsn": "string_value",
                "gst_rate": "string_value",
                "cess": "string_value",
                "gst_category": "string_value",
                "nil_rated": "string_value",
                "non_gst_item": "string_value",
                "for_web": "string_value",
                "manufacturer": "string_value",
                "location": "string_value",
                "schedule": "string_value",
                "main_image1": 30,
                "main_image2": 30,
                "detail": "string_value",
                "ean_bar_code": "string_value",
                "no_med_rem": "string_value",
                "linked_item_store": "string_value",
                "qty": "string_value",
                "medicine_type": "string_value",
                "manufacture_date": "string_value",
                "expiry_date": "string_value",
                "batch_number": "string_value",
                "storage_conditions": 30,
                "POST": "string_value"
              }
            }
          }
        }
      }
    },
    "/inventory": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "summary": "Inventory - /inventory",
        "responses": {
          "200": {
            "description": "Successful Operation"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/dropdown/medicines": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "summary": "Inventory - /dropdown/medicines",
        "responses": {
          "200": {
            "description": "Successful Operation"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}